+++ ambientika-local-control/src/models/device-broadcast-status.model.ts
@@ -3,12 +3,14 @@ export class DeviceBroadcastStatus {
     zoneIndex: number;
     fanMode: string;
     fanStatus: string;
+    houseId: number | undefined;
 
 
-    constructor(serialNumber: string | undefined, zoneIndex: number, fanMode: string, fanStatus: string) {
+    constructor(serialNumber: string | undefined, zoneIndex: number, fanMode: string, fanStatus: string, houseId?: number) {
         this.serialNumber = serialNumber;
         this.zoneIndex = zoneIndex;
         this.fanMode = fanMode;
         this.fanStatus = fanStatus;
+        this.houseId = houseId;
     }
 }